src/production/requirements.txt
pandas==1.3.5
numpy==1.21.5
openai==1.6.1
streamlit==1.21.0
python-dotenv==1.0.0

